requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
captcha==0.5.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.1-rc1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-countries==7.6.1
django-debug-toolbar==4.4.6
django-jazzmin==3.0.0
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
exceptiongroup==1.2.2
gunicorn==22.0.0
hiredis==3.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
packaging==24.1
phonenumbers==8.13.27
pillow==10.3.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.3.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.8
referencing==0.32.1
rpds-py==0.16.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
